Block Girafarig that can Mega Evolve into Mega Girafarig from Evolving
When the hidden FLINT is activated, even if a Girafarig gains the ability to Mega Evolve into the Q.Qore Mega Girafarig by holding the Girafarigite, it can evolve into Flagona. This is not the intended behavior. All Pokemon that will have the ability to Mega Evolve shouldn't be able to evolve, similar to Pikachu, Eevee and Meowth that can gigantamax