Begin Ι2Ε (#23)
* Begin Ι2Ε

All stabilization project edits that edit scripts and PBS will go there.

Once this PR gets merged, Qora Qore ΙΨ will be released

* Update Pokémon Cry Style to Pokémon Cry Sounds, thanks Surpalina

To reflect that the Classic Cries no longer exist. The dreamy prograda also has a "Prograda" option

* Tweak certain descriptions to fit in 3 lines of text while Didact Gothic is used as main font

A few of the info has been put into another bullet sub-entry in the Q.Qore Data Information file but within the same bullet

* Remove Drop Shadow from Glourk Header Titletext to match the rest of the static text

As a result, certain text in light mode became dark as it will produce lower contrast without shadow text
This makes the Glourk Header Titlebar text similar (But not the same) as the static text in Glourk Headers

* Reposition Wikicities slightly

As sprites were slightly updated

* Redesign Pokemon Mart and Move Relearner Screens to use Glourk 5 Headers

Summary screens cannot be adjusted as those need to be fullscreen because some pages use whole screen height, similar to Purify Chamber

* Redesign Hall of Fame Screen and a few changes

For HoF and Bag Screens, they take full advantage of the 640x384 screen size
For the Loadscreen, removed the button reference

* Increase Size of Move Details Box

For Summary Screens, a design change based on the Move Relearner Screen variant was also done in addition to the size increase

* Allow getting Internal Channel colorization by editing game.ini

You can get this by changing the Channel property in the Qortex Section of Game.ini to either "Internal" or "Upgrade Wizard" but it is not recommended to do it so

* Bag and Pokédex Screen changes

For the formeer, slider is now covering the whole screen vertically where it can
For the latter, the Search UI is no longer Dark with Light Canvas in Light Mode, keeping the former color scheme for the Dark Mode albeit with dark canvas

* Support changning Text Color Scheme

Text Skin Color Scheme changes what colors you get on text skins with \m, \f, \pg, \pog and \c[1] through \c[8]. It does not change anything else
In addition, text shadow on male and female text is much darker on a Dark Skin to further improve contrast

* Male Text Color on Colors Color Scheme should be dark in light mode

* Pokedex Seen Graphic must follow System Theme like the Owned ones

* Make Hall of Fame Screen Theme Aware

Note that more changes were required to be done by adding extra code

* Transform the HoF screen bottom part to a Glourk Footer

This adds a top border of 4px and drops shadow text as Glourk Headers and Footers no longer have shadow text in it but this does not drop entirely from the game the Shadowed text
No other changes were done in this additional CU

* Tweak how Egg and Remote Box sprites are being gathered.

Now support New6, Gender Differences, Shinies, Backs and Shadowness in addition to base sprite and form differences
This also gives the foundation of making gender differences visible to remove boxes

For female eggs, it is "000eggf".
For shiny remote bxoes, it is "000remoteboxs"
For female remote boxes, it is "000remoteboxf"
For alternative form remote boxes, it is still as before like "000remotebox_1"
For female crack files, it is "000eggCracksf"

* Alter and Mega Channel must be able to learn Lick Lock and Herbotopia via Breeding

Combination of a Male Smeargle knowning them and a Female Mega Channel knowing those moves is also one of the ways to get those moves into Alter or Mega Channel
Remember, breeding with Mega Channel has a 50% chance of getting an Alter Channel egg

Egg sprites for Elemental Bobojido Pokmemon, Google Play Store, Xbox, Bing, Azure, Visual Studio, Bulbagarden, Discord, Alter Channel and Mega Channel were also added

* Season-aware Day and Night System checks

Along with certain fixes and tweaks
Also fixes a softlocking issue when exiting the Settings Screen without entering in one of the subpages first. Said screen's code was changed as well, to be consistent with the other screens, leaving the Debug screen code the ones without this treatment

* Put AztecCroc into credits list

Needed as Q.Qore uses the Blank Plate Icon