Merged Junior and Basic boxes into one
This is the Tutorial Box combining the best of the two levels.
Also, you can now set the Box level directly on the box instance instead of being hardcoded for specific ones.
Lastly, the Box Name will be shown on the titlebar too, but on Milenial Boxes, the name of the last standard box level is shown instead with the actual name being seen once in the X Challenge Ahead! message and in the summary screen
Classic Box now has a duration of 10 days instead of 7 days
Bumps delta to .96